pyproject.toml
[project]
name = "telegram-rate-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Vitaliy Pavlov",email = "v-2841@yandex.ru"}
]
readme = "README.md"
requires-python = "3.13.*"
dependencies = [
    "python-telegram-bot[rate-limiter] (>=22.4,<23.0)",
    "aiohttp (>=3.12.15,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.poetry]
package-mode = false
